begin modularizing redux
