+++ packages/common/src/index.ts
@@ -1,34 +1,13 @@
 import * as redux from "redux"
-import { createStore, applyMiddleware, Middleware } from "redux"
+import { applyMiddleware, createStore, Middleware } from "redux"
 import { createLogger } from "redux-logger"
+import { actions, counter } from "./redux/modules/counter"
 
-export interface IncrementAction {
-    readonly type: string
-}
-
-export interface IDecrementAction {
-    readonly type: string
-}
-
-export interface IResetAction {
-    readonly type: string
-}
+import createSagaMiddleware from "redux-saga"
 
 export interface IAppState {
     readonly counter: number
 }
-
-const increment = (): IncrementAction => ({
-    type: "INCREMENT",
-})
-
-const decrement = (): IDecrementAction => ({
-    type: "DECREMENT",
-})
-
-const reset = (): IResetAction => ({
-    type: "RESET",
-})
 
 const appState: IAppState = {
     counter: 0,
@@ -36,36 +15,18 @@ const appState: IAppState = {
 
 const INITIAL_STATE: IAppState = appState
 
-const counter = (
-    state = INITIAL_STATE.counter,
-    action: IncrementAction | IDecrementAction,
-): number => {
-    switch (action.type) {
-        case "INCREMENT":
-            return state + 1
-        case "DECREMENT":
-            return state - 1
-        case "RESET":
-            return INITIAL_STATE.counter
-        default:
-            return state
-    }
-}
-
 const appReducer = redux.combineReducers({ counter })
 const logger = createLogger({ stateTransformer: (state: IAppState) => state })
-const middleware: ReadonlyArray<Middleware> = [logger]
+const sagaMiddleware = createSagaMiddleware()
+const middleware: ReadonlyArray<Middleware> = [sagaMiddleware, logger]
 
-const createStoreWithMiddleware = () => {
-    return createStore(
-        appReducer,
-        INITIAL_STATE,
-        applyMiddleware(...middleware),
-    )
-}
+const createStoreWithMiddleware = createStore(
+    appReducer,
+    INITIAL_STATE,
+    applyMiddleware(...middleware),
+)
 
-const store = createStoreWithMiddleware()
+const store = createStoreWithMiddleware
+// sagaMiddleware.run(Sagas.root)
 
-const actions = { increment, decrement, reset }
-
-export { actions, appState, store }
+export { actions, appState, store, INITIAL_STATE }
